MVsB5_BT_Audio_SDK_V0.2.8 - *.c *.h

+++ BT_Audio_APP/app_framework/audio_engine/recorder_service.h
@@ -44,7 +44,7 @@ extern "C"{
 
 uint16_t RecFileIndex(char *string);
 void IntToStrMP3Name(char *string, uint16_t number);
-void RecServierMsg(uint16_t *msg);
+void RecServierMsg(uint32_t *msg);
 xTaskHandle GetMediaRecorderTaskHandle(void);
 bool MediaRecorderServiceDeinit(void);
